If no third party information exists, don't run delete logic (#117)

+++ webapp/src/main/java/com/box/l10n/mojito/service/screenshot/ScreenshotService.java
@@ -412,10 +412,11 @@ public class ScreenshotService {
               .getRepository()
               .getName();
 
-      String projectId = thirdPartySyncJobs.get(repository).getThirdPartyProjectId();
-
-      thirdPartyService.removeImage(projectId, thirdPartyScreenshot.getThirdPartyId());
-      thirdPartyScreenshotRepository.deleteById(thirdPartyScreenshot.getId());
+      if (thirdPartySyncJobs.containsKey(repository)) {
+        String projectId = thirdPartySyncJobs.get(repository).getThirdPartyProjectId();
+        thirdPartyService.removeImage(projectId, thirdPartyScreenshot.getThirdPartyId());
+        thirdPartyScreenshotRepository.deleteById(thirdPartyScreenshot.getId());
+      }
     }
 
     screenshotTextUnitRepository.deleteAllByScreenshot_Id(id);
